Update capybara matcher
The label is always on the page, so it grabs the label and then checks to see if the filters match.  They will not match if the code hasn't finished executing.  The filters are not part of the timeout cycle, so the previous selector may or may not work depending on whether the code has finished executing.