Add ASpace uploader headers file
When uploading to ArchivesSpace, special headers have to be used
in the excel file, including a row of cells with mysterious
invisible formatting that is required for the upload to succeed.

This adds the relevant headers so that their mysterious blank cells
can be copied after uploader sheets are regenerated, to make them
successfully upload.

See the source_data/README for more information.